finish conversion of project to typescript
