+++ package.json
@@ -18,7 +18,7 @@
     "react": "^18.1.0",
     "react-datepicker": "^4.8.0",
     "react-dom": "^18.1.0",
-    "react-hook-form": "^7.31.3",
+    "react-hook-form": "^7.48.2",
     "react-redux": "^8.0.1",
     "react-router": "^6.4.1",
     "react-router-dom": "^6.3.0",
