+++ src/styles/App.css
@@ -1,8 +1,4 @@
 
-/*@font-face {*/
-/*  font-display: block;*/
-/*}*/
-
 * {
   box-sizing: border-box;
 }
@@ -14,8 +10,6 @@ html {
   overflow-y: scroll;
 
 }
-
-
 
 @media (min-width: 325px) {
   html {
