+++ src/styles/Themes.ts
@@ -1,0 +1,58 @@
+export const lightTheme = {
+  body: "#F2F2F2",
+  text: "hsl(200, 15%, 8%)",
+  textPlain: "black",
+  toggleBorder: "#FFF",
+  background: "#FFFFFF",
+  font: "League Spartan",
+  elements: "hsl(0, 0%, 100%)",
+  outline: "#7C5DFA",
+  greyText: "#888EB0",
+  formBackground: "#FFFFFF",
+  headerBackground: "#373B53",
+  formFieldOutline: "#DFE3FA",
+  formFieldOutlineFocus: "#9277FF",
+  newButton: "#7C5DFA",
+  newButtonHover: "#9277FF",
+  editButton: "#F9FAFE",
+  editButtonHover: "#DFE3FA",
+  newItemText: "#7E88C3",
+  inputBackgroundColor: "#FFFFFF",
+  draftColor: "#373B53",
+  draftBackgroundColor: "rgba(55,59,83, 0.06)",
+  filterShadow: "0px 10px 20px rgba(72, 84, 159, 0.25)",
+  dateText: "#0C0E16",
+  amountDueBackground: "#373B53",
+  itemListFormBackground: "white",
+  paymentTermsBackground: "white",
+  paymentOptionBorder: "#DFE3FA",
+};
+export const darkTheme = {
+  body: "#141625",
+  text: "hsl(0, 0%, 100%)",
+  textPlain: "#FFFFFF",
+  headerBackground: "#1E2139",
+  toggleBorder: "#6B8096",
+  background: "#1E2139",
+  elements: "hsl(209, 23%, 22%)",
+  outline: "#7C5DFA",
+  greyText: "#DFE3FA",
+  formBackground: "#141625",
+  formFieldOutline: "#252945",
+  formFieldOutlineFocus: "#9277FF",
+  newButton: "#7C5DFA",
+  newButtonHover: "#9277FF",
+  editButton: "#252945",
+  editButtonHover: "#FFFFFF",
+  newItemText: "#888EB0",
+  inputBackgroundColor: "#1E2139",
+  draftColor: "#DFE3FA",
+  draftBackgroundColor: "rgba(151, 151, 151, 0.06)",
+  filterShadow: "0px 10px 20px rgba(0, 0, 0, 0.25)",
+  dateText: "#DFE3FA",
+  amountDueBackground: "#0C0E16",
+  itemListFormBackground: "#141625",
+  paymentTermsBackground: "#252945",
+  paymentOptionBorder: "#1E2139",
+};
+
